feat: Upload 의존성 추가

+++ src/main/java/com/prgrms/p2p/domain/course/entity/CoursePlace.java
@@ -108,7 +108,7 @@ public class CoursePlace extends BaseEntity {
     this.place = place;
   }
 
-  private void setIndex(Long index) {
+  private void setIndex(Integer index) {
     this.index = index;
   }
 }
